feat : add system information table and add default configures - fix : add field and change field type in a few table : staking, user_staking

+++ apps/engine/schemas/financial_transactions.py
@@ -15,7 +15,7 @@ class FinancialTransactions(models.Model):
         max_length=32,
         choices=TypeTransactionEnum.choices(),
     )
-    amount = models.DecimalField(max_digits=15, decimal_places=2)
+    amount = models.PositiveIntegerField()
     tx_hash = models.CharField(max_length=256)
 
     created_at = models.DateTimeField(auto_now_add=True)
